+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2018-08-18
+### Added
+- Datetime value objects
+- Duration value objects
+
 ## [1.1.2] - 2018-08-05
 ### Changed
 - Re-release of the package